+++ ui/src/app/components/shopping-cart/shopping-cart.component.spec.ts
@@ -1,8 +1,8 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from "@angular/core/testing";
 
-import { ShoppingCartComponent } from './shopping-cart.component';
+import { ShoppingCartComponent } from "./shopping-cart.component";
 
-describe('ShoppingCartComponent', () => {
+describe("ShoppingCartComponent", () => {
   let component: ShoppingCartComponent;
   let fixture: ComponentFixture<ShoppingCartComponent>;
 
@@ -15,7 +15,7 @@ describe('ShoppingCartComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should create', () => {
+  it("should create", () => {
     expect(component).toBeTruthy();
   });
 });